use modified version of ui.js
This way, we get a very lightweight console, because we avoid loading
all ExtJS classes. Also it is fully functional on mobile devices now.

I removed the websockify code because we implement websockets in HTTPServer.pm.

The html code moved to pve-manager, because we need to include CSRF token and
gettext catalogs.