Replace 'migrating' info strings with more generic 'connecting' messages
This fixes the problem that a restarted container was seen as migrating
in the UI.

Background:
The check for migration happens on a 'disconnected' event, but can very well
come for other reasons, like entering reboot on the LXC console.
Since we have no idea about what have caused the disconnection, it is
better to display a more generic message.