parse 'already running' error on guest start
Starting an already running container results in an sync-error
instead of a task that fails async. So, to have similar behaviour
between VMs and CTs, check any request error for the start command
and silent the 'already running' one.

Put this change in the first downstream patch and rebase the rest of
them

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>