suppress unauthorized error from status query for VM.Console only user
We now query the guest status before starting the console to provide
a start button if it's not yet running, but that requires VM.Audit
privileges, so more than just VM.Console on the guest.

So, ignore a `403 Forbidden` response, as that just means we cannot
query the VM info to see if it's running, rather just continue
connecting.

See: https://forum.proxmox.com/threads/116176/
Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
 [ T: reworked commit message ]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>